Allow threshold == set size and add tests
